requirements.txt
pandas>=2.0
sqlalchemy>=2.0
python-dotenv>=1.0
scikit-learn>=1.3
lightgbm>=4.0
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
jupyterlab>=4.0
rapidfuzz>=3.0
